@@ -65,7 +65,7 @@ class DailyForecastsWidget extends ConsumerWidget {
                         height: 6.h,
                       ),
                       Icon(
-                        FontAwesomeIcons.tint,
+                        FontAwesomeIcons.droplet,
                         color: Theme.of(context).textTheme.subtitle2!.color,
                         size: kIconSize(context),
                       ),
